+++ src/scorer/feature/retoots_in_feed_scorer.ts
@@ -1,0 +1,17 @@
+/*
+ * Score how many times a toot has been retooted by other accounts in the feed.
+ */
+import FeatureScorer from '../feature_scorer';
+import Toot from '../../api/objects/toot';
+import { WeightName } from "../../types";
+
+
+export default class RetootsInFeedScorer extends FeatureScorer {
+    constructor() {
+        super(WeightName.RETOOTED_IN_FEED);
+    }
+
+    async _score(toot: Toot) {
+        return toot.reblogsBy.length;
+    }
+};