+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* RetootsInFeedScorer can use Toot.reblogsBy.length to score (and therefore doesn't need to be a FeedScorer)
 
 ### v0.20.7
 * Consolidate all `Toot` post processing to `Toot.setDependentProperties()` method